.github/workflows/build: Run by cron schedule every day
As we're testing a submodule of bigger project here, it's not enough to
trigger CI on just changes to the submodule, as the build issues can
arise due to the changes in the main project. As we can't easily
track changes to the main project, just perform daily builds against
it.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>