.github/workflows/build.yml: Pin OS to ubuntu-20.04.
Before this change, we effectively used ubuntu-18.04.

The specific situation which prompts this change is that Python
"cryptography" package got a major internal upgrade to use Rust code.
At the same time, its binary (.whl) packages were upgraded to modern
version, which cannot be installed with (pretty old by now) "pip3"
version shipped with Ubuntu 18.04.

Finally, "ubuntu-latest" (as used now) is anyway scheduled to be
switched to ubuntu-20.04, so we're just taking steps ahead, and
follow the best practice to pin (system-level) dependencies.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>