pkcs11: Fix a memory leak in `C_FindObjects`
The corePKCS11
(https://github.com/FreeRTOS/corePKCS11/blob/main/source/core_pkcs11.c)
expects that allocated memory is freed by the API `C_FindObjects` in
the case of error in finding the requested object. This expectation was
not met, which led to a memory leak when finding the requested object
fails. Release the allocated memory, if returning an error from
`C_FindObjects`.

Signed-off-by: Devaraj Ranganna <devaraj.ranganna@arm.com>